A giant update 0.2.0. Chatting context saving. Any other improvements.
