+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pts_gem-chat"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2024"
 
 [dependencies]
